Move HttpAdapter to own file
